[Rust][refactor](executor, import_obj, statistics): Refactored the unit tests.
