+++ bindings/rust/wasmedge-sys/src/executor.rs
@@ -140,7 +140,7 @@ impl Executor {
         Ok(self)
     }
 
-    /// Invokes a WASM function in the anonymous [`Module`].
+    /// Invokes a WASM function in the anonymous [`Module`], and returns the results.
     ///
     /// After instantiating a WasmEdge [`Module`], the [`Module`] is registered as an
     /// anonymous module in the [`Store`]; then, you can repeatedly call this function
@@ -165,7 +165,7 @@ impl Executor {
         store: &Store,
         func_name: impl AsRef<str>,
         params: impl IntoIterator<Item = Value>,
-    ) -> WasmEdgeResult<impl Iterator<Item = Value>> {
+    ) -> WasmEdgeResult<Vec<Value>> {
         let raw_params = params
             .into_iter()
             .map(wasmedge::WasmEdge_Value::from)
@@ -189,10 +189,10 @@ impl Executor {
             returns.set_len(returns_len);
         }
 
-        Ok(returns.into_iter().map(Into::into))
+        Ok(returns.into_iter().map(Into::into).collect::<Vec<_>>())
     }
 
-    /// Invokes a registered WASM function by its module name and function name.
+    /// Invokes a registered WASM function by its module name and function name, and returns the results.
     ///
     /// # Arguments
     ///
@@ -214,7 +214,7 @@ impl Executor {
         mod_name: impl AsRef<str>,
         func_name: impl AsRef<str>,
         params: impl IntoIterator<Item = Value>,
-    ) -> WasmEdgeResult<impl Iterator<Item = Value>> {
+    ) -> WasmEdgeResult<Vec<Value>> {
         let raw_params = params
             .into_iter()
             .map(wasmedge::WasmEdge_Value::from)
@@ -243,7 +243,7 @@ impl Executor {
             returns.set_len(returns_len);
         }
 
-        Ok(returns.into_iter().map(Into::into))
+        Ok(returns.into_iter().map(Into::into).collect::<Vec<_>>())
     }
 }
 impl Drop for Executor {
@@ -251,5 +251,32 @@ impl Drop for Executor {
         if !self.ctx.is_null() {
             unsafe { wasmedge::WasmEdge_ExecutorDelete(self.ctx) }
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{Config, Statistics};
+
+    #[test]
+    fn test_executor_create() {
+        let result = Executor::create(None, None);
+        assert!(result.is_ok());
+
+        let result = Config::create();
+        assert!(result.is_ok());
+        let config = result.unwrap();
+        let result = Executor::create(Some(&config), None);
+        assert!(result.is_ok());
+
+        let result = Statistics::create();
+        assert!(result.is_ok());
+        let mut stat = result.unwrap();
+        let result = Executor::create(None, Some(&mut stat));
+        assert!(result.is_ok());
+
+        let result = Executor::create(Some(&config), Some(&mut stat));
+        assert!(result.is_ok());
     }
 }
