+++ bindings/rust/wasmedge-sys/src/statistics.rs
@@ -90,75 +90,3 @@ impl Drop for Statistics {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{Config, Executor, Loader, Validator};
-
-    #[test]
-    fn test_stat() {
-        // create a Config context
-        let result = Config::create();
-        assert!(result.is_ok());
-        let config = result.unwrap();
-        // enable Statistics
-        let config = config
-            .count_instructions(true)
-            .measure_time(true)
-            .measure_cost(true);
-
-        // load module from a wasm file
-        let result = Loader::create(Some(&config));
-        assert!(result.is_ok());
-        let loader = result.unwrap();
-        let path =
-            std::path::PathBuf::from(env!("WASMEDGE_DIR")).join("test/api/apiTestData/test.wasm");
-        let result = loader.from_file(path);
-        assert!(result.is_ok());
-        let module = result.unwrap();
-
-        // validate module
-        let result = Validator::create(Some(&config));
-        assert!(result.is_ok());
-        let validator = result.unwrap();
-        let result = validator.validate(&module);
-        assert!(result.is_ok());
-
-        // create a Statistics context
-        let result = Statistics::create();
-        assert!(result.is_ok());
-        let mut stat = result.unwrap();
-
-        // set cost table
-        let mut cost_table = vec![20u64; 512];
-        stat.set_cost_table(&mut cost_table);
-
-        // set cost limit
-        stat.set_cost_limit(100_000_000_000_000);
-
-        // create an Executor context
-        let result = Executor::create(Some(&config), Some(&mut stat));
-        assert!(result.is_ok());
-        // let executor = result.unwrap();
-
-        // register an ImportObj module
-        // let import_obj = create_extern_module("extern", false);
-
-        // todo: Add more test code
-    }
-
-    // fn create_extern_module(name: impl AsRef<str>, wrapped: bool) -> ImportObj {
-    //     // create an ImportObj module
-    //     let result = ImportObj::create(name);
-    //     assert!(result.is_ok());
-    //     let import_obj = result.unwrap();
-
-    //     // // create a FuncType instance
-    //     // let params = vec![ValType::ExternRef, ValType::I32];
-    //     // let returns = vec![ValType::I32];
-    //     // let result = FuncType::create(params, returns);
-
-    //     import_obj
-    // }
-}
